replaced TrajectoryFlowRegionQuery with FlowRegionConstraint
